Adding training scripts
