--- docker/Dockerfile.noetic
@@ -1,43 +1,0 @@
-FROM nvcr.io/nvidia/pytorch:21.11-py3
-
-# Copyright (c) 2020, NVIDIA CORPORATION. All rights reserved.
-# Full license terms provided in LICENSE.md file.
-# To build:
-# docker build -t nvidia-dope:noetic-v1 -f Dockerfile.noetic ..
-
-ENV HOME /root
-ENV DEBIAN_FRONTEND=noninteractive
-
-# Install system and development components
-RUN apt-get update && apt-get -y --no-install-recommends install \
-    apt-utils \
-    software-properties-common \
-    build-essential \
-    cmake \
-    git \
-    python3-pip \
-    libxext6 \
-    libx11-6 \
-    libglvnd0 \
-    libgl1 \
-    libglx0 \
-    libegl1 \
-    freeglut3-dev \
-    && apt-get -y autoremove \
-    && apt-get clean
-
-# pip install required Python packages
-COPY requirements.txt ${HOME}
-RUN python3 -m pip install --no-cache-dir -r ${HOME}/requirements.txt
-
-COPY scripts/ ${HOME}/scripts/
- 
-ENV DISPLAY :0
-ENV NVIDIA_VISIBLE_DEVICES all
-ENV NVIDIA_DRIVER_CAPABILITIES graphics,utility,compute
-ENV TERM=xterm
-ENV PYTHONPATH "${PYTHONPATH}:/usr/local/lib:/workspace/dope/sr"
-# Some QT-Apps don't show controls without this
-ENV QT_X11_NO_MITSHM 1
-
-# ENTRYPOINT ["python3", "./scripts/train.py"]
