+++ readme.md
@@ -9,6 +9,8 @@ This is the official DOPE ROS package for detection and 6-DoF pose estimation of
 ![DOPE Objects](dope_objects.png)
 
 ## Updates
+2021/12/13 - Added a NViSII script to generate synthetic data for training DOPE. See this [readme](https://github.com/NVlabs/Deep_Object_Pose/tree/master/scripts/nvisii_data_gen) for more details. We also added the update training and inference (without ROS) scripts for the NViSII paper [here](https://github.com/NVlabs/Deep_Object_Pose/tree/master/scripts/train2). 
+
 2021/10/20 - Added ROS2 Foxy inference support through [Isaac ROS DOPE package](https://github.com/NVIDIA-ISAAC-ROS/isaac_ros_dnn_inference/tree/main/isaac_ros_dope) for Jetson and x86+CUDA-capable GPU.
 
 2021/08/07 - Added publishing belief maps. Thank you to Martin Günther. 
@@ -19,7 +21,7 @@ This is the official DOPE ROS package for detection and 6-DoF pose estimation of
 
 2020/16/03 - Added a wiki (thanks to [@saratrajput](https://github.com/saratrajput)) 
 
-2019/03/07 - ROS interface update (thanks to Martin Günther)
+<!-- 2019/03/07 - ROS interface update (thanks to Martin Günther) -->
 
 <!-- 2019/11/06 - Added bleach YCB weights  -->
