+++ scripts/hyperparameters.json
@@ -1,4 +1,4 @@
 {
-    "obj" : "010_potted_meat_can_16k",
+    "obj" : "meat",
     "gpus" : "0"
 }
